\name{add.to.branches.and.nodes}
\alias{add.to.branches.and.nodes}
\title{Add tip to branches and nodes of the tree}
\usage{
add.to.branches.and.nodes(tree,tip.name)
}
\arguments{
  \item{tree}{phylogeny.}
  \item{tip.name}{name.}
}
\description{
  Internal function for \code{\link{allFurcTrees}}.
}
\details{
  Do not use unless you know what you're doing.
}
\value{
  a list of trees as an object of class \code{"multiPhylo"}.
}
\references{
}
\author{Liam Revell \email{liam.revell@umb.edu}}
\seealso{
}
\examples{
}
\keyword{phylogenetics}
