\name{anoletree}
\alias{anoletree}
\alias{salamanders}
\title{Phylogenetic datasets}
\description{
	\code{anoletree} is a phylogeny of Greater Antillean anole species with a mapped discrete character - \emph{ecomorph class}. Data and tree are from Mahler et al. (2010).
	
	\code{salamanders} is a phylogeny of \emph{Plethodon} salamanders from Highton and Larson (1979). According to Wikipedia, the genus \emph{Plethodon} contains 55 species in total.
}
\usage{
data(anoletree)
data(salamanders)
}
\format{
	The data are stored as a modified object of class \code{"simmap"} with a mapped discrete character. (E.g., see \code{\link{read.simmap}}.)
	
	Note that the mapping of the discrete character is (in this case) arbitrary.
}
\source{
	Highton, R., and A. Larson. (1979) The genetic relationships of the salamanders of the genus \emph{Plethodon}. \emph{Systematic Zoology}, \bold{28}, 579-599.

	Mahler, D. L, L. J. Revell, R. E. Glor, and J. B. Losos. (2010) Ecological opportunity and the rate of morphological evolution in the diversification of Greater Antillean anoles. \emph{Evolution}, \bold{64}, 2731-2745.
}
\keyword{datasets}
