\name{ave.rates}
\alias{ave.rates}
\title{Compute weighted average evolutionary rates}
\usage{
ave.rates(tree,shift,tips,sig1,sig2,ave.shift,showTree=TRUE)
}
\arguments{
  \item{tree}{phylogeny.}
  \item{shift}{.}
  \item{tips}{.}
  \item{sig1}{.}
  \item{sig2}{.}
  \item{ave.shift}{.}
  \item{showTree}{.}
}
\description{
  Internal function for \code{\link{posterior.evolrate}}.
}
\details{
  Do not use unless you know what you're doing.
}
\value{
 a list with the rates.
}
\references{
}
\author{Liam Revell \email{liam.revell@umb.edu}}
\seealso{
}
\examples{
}
\keyword{phylogenetics}
