\name{branch.and.bound}
\alias{branch.and.bound}
\title{Perform branch & bound parsimony search}
\usage{
branch.and.bound(data,tree)
}
\arguments{
  \item{data}{data.}
  \item{tree}{phylogeny.}
}
\description{
  Internal function for \code{\link{exhaustiveMP}}.
}
\details{
  Do not use unless you know what you're doing.
}
\value{
  a tree or list of trees.
}
\references{
}
\author{Liam Revell \email{liam.revell@umb.edu}}
\seealso{
}
\examples{
}
\keyword{phylogenetics}
