\name{compute.ancestor.nodes}
\alias{compute.ancestor.nodes}
\title{Compute all ancestral nodes}
\usage{
compute.ancestor.nodes(tree)
}
\arguments{
  \item{tree}{.}
}
\description{
  Internal function for \code{\link{optim.phylo.ls}}.
}
\details{
  Do not use unless you know what you're doing.
}
\value{
  a matrix.
}
\references{
}
\author{Liam Revell \email{liam.revell@umb.edu}}
\seealso{
}
\examples{
}
\keyword{phylogenetics}
