\name{exhaustive.search}
\alias{exhaustive.search}
\title{Exhaustive MP search}
\usage{
exhaustive.search(data)
}
\arguments{
  \item{data}{.}
}
\description{
  Internal function for \code{\link{exhaustiveMP}}.
}
\details{
  Do not use unless you know what you're doing.
}
\value{
  a tree or list of trees.
}
\references{
}
\author{Liam Revell \email{liam.revell@umb.edu}}
\seealso{
}
\examples{
}
\keyword{phylogenetics}
