\name{expm}
\alias{expm}
\title{Matrix exponential}
\usage{
expm(Y)
}
\arguments{
  \item{Y}{a matrix.}
}
\description{
  Wrapper for \code{\link{MatrixExp}} that retains row/column names.
}
\details{
}
\value{
  a matrix.
}
\references{
}
\author{Liam Revell \email{liam.revell@umb.edu}}
\seealso{
}
\examples{
}
\keyword{math}
