\name{get.treepos}
\alias{get.treepos}
\title{Internal function}
\usage{
get.treepos(message=TRUE, ...)
}
\arguments{
	\item{message}{argument.}
	\item{...}{optional arguments.}
}
\description{
	Internally used function.
}
\value{
	A list.
}
\references{
	Revell, L. J. (2012) phytools: An R package for phylogenetic comparative biology (and other things). \emph{Methods Ecol. Evol.}, \bold{3}, 217-223.
}
\author{Liam Revell \email{liam.revell@umb.edu}}
\keyword{statistics}
\keyword{utilities}
