\name{likelihood.lambda}
\alias{likelihood.lambda}
\title{Likelihood of lambda transformation}
\usage{
likelihood.lambda(lambda,y,X,C)
}
\arguments{
  \item{lambda}{.}
  \item{y}{.}
  \item{X}{.}
  \item{C}{.}
}
\description{
  Internal function for \code{\link{phyl.resid}}.
}
\details{
  Do not use unless you know what you're doing.
}
\value{
  likelihood.
}
\references{
}
\author{Liam Revell \email{liam.revell@umb.edu}}
\seealso{
}
\examples{
}
\keyword{phylogenetics}
