\name{likelihood}
\alias{likelihood}
\title{Likelihood of evolutionary vcv matrix, for a particular lambda}
\usage{
likelihood(theta,X,C)
}
\arguments{
  \item{theta}{.}
  \item{X}{.}
  \item{C}{.}
}
\description{
  Internal function for \code{\link{phyl.pca}}.
}
\details{
  Do not use unless you know what you're doing.
}
\value{
  likelihood.
}
\references{
}
\author{Liam Revell \email{liam.revell@umb.edu}}
\seealso{
}
\examples{
}
\keyword{phylogenetics}
