\name{ls.tree}
\alias{ls.tree}
\title{Least squares branch lengths for a given tree}
\usage{
ls.tree(tree,D)
}
\arguments{
  \item{tree}{.}
  \item{D}{.}
}
\description{
  Internal function for \code{\link{optim.phylo.ls}}.
}
\details{
  Do not use unless you know what you're doing.
}
\value{
  a tree with branch lengths.
}
\references{
}
\author{Liam Revell \email{liam.revell@umb.edu}}
\seealso{
}
\examples{
}
\keyword{phylogenetics}
