\name{mapped.states}
\alias{mapped.states}
\title{Returns a vector, matrix, or list of the mapped states on a tree or set of trees}
\usage{
mapped.states(tree, ...)
}
\arguments{
	\item{tree}{a single tree or a set of trees as an object of class \code{"simmap"} or \code{"multiSimmap"}, respectively.}
	\item{...}{optional arguments.}
}
\description{
	Computes and orders a vector, matrix, or list of the unique mapped states on a tree or state of trees of class \code{"simmap"} or \code{"multiSimmap"}.
}
\value{
	A vector, matrix, or list.
}
\references{
	Revell, L. J. (2012) phytools: An R package for phylogenetic comparative biology (and other things). \emph{Methods Ecol. Evol.}, \bold{3}, 217-223.
}
\author{Liam Revell \email{liam.revell@umb.edu}}
\keyword{comparative method}
\keyword{phylogenetics}
\keyword{utilities}
