\name{phyl.vcv}
\alias{phyl.vcv}
\title{Internal function}
\usage{
phyl.vcv(X,C,lambda)
}
\arguments{
  \item{lambda}{.}
  \item{X}{.}
  \item{C}{.}
}
\description{
  Internal function for \code{\link{phyl.pca}}.
}
\details{
  Do not use unless you know what you're doing.
}
\value{
  matrix.
}
\references{
}
\author{Liam Revell \email{liam.revell@umb.edu}}
\seealso{
}
\examples{
}
\keyword{phylogenetics}
