\name{phyloDesign}
\alias{phyloDesign}
\title{Compute design matrix for least squares analyses}
\usage{
phyloDesign(tree)
}
\arguments{
  \item{tree}{.}
}
\description{
  Internal function for \code{\link{optim.phylo.ls}}.
}
\details{
  Do not use unless you know what you're doing.
}
\value{
  a matrix.
}
\references{
}
\author{Liam Revell \email{liam.revell@umb.edu}}
\seealso{
}
\examples{
}
\keyword{phylogenetics}
