\name{phytools-package}
\alias{phytools-package}
\alias{phytools}
\docType{package}
\title{
phytools: Phylogenetic Tools for comparative biology (and other things)
}
\description{
  \pkg{phytools} provides functions for phylogenetic comparative biology; as well as several other functions for tree inference, manipulation, and analysis that are not implemented in other R packages.

  The complete list of functions can be displayed with
  \code{library(help = phytools)}.

  More information on \pkg{phytools} can be found at \url{http://anolis.oeb.harvard.edu/~liam/R-phylogenetics/}.
}

\author{
  Liam J. Revell

  Maintainer: Liam J. Revell <liam.revell@umb.edu>
}
\references{
    Revell, L. J. (2011). phytools: Phylogenetic Tools for Comparative Biology (and other things). http://anolis.oeb.harvard.edu/~liam/R-phylogenetics
}
\keyword{package}
