\name{posterior.evolrate}
\alias{posterior.evolrate}
\title{Analysis of the posterior sample from evol.rate.mcmc}
\usage{
posterior.evolrate(tree, ave.shift, mcmc, tips, showTree=FALSE)
}
\arguments{
	\item{tree}{a phylogenetic tree in \code{"phylo"} format.}
	\item{ave.shift}{mean or median shift-point from the posterior sample (see \code{\link{minSplit}}.}
	\item{mcmc}{matrix \code{$mcmc} from \code{evol.rate.mcmc} (probably with burn-in excluded).}
	\item{tips}{list of tips in state \eqn{\sigma_1^2} for each sampled generation of MCMC.}
	\item{showTree}{optional logical value indicating whether or not to plot the stretched and shrunken tree generated by the pre-processing algorithm implemented in this function (default is \code{FALSE}).}
}
\description{
	Analyzes posterior sample from \code{\link{evol.rate.mcmc}}.
}
\details{
	This function takes a phylogenetic tree, an average split position, and a raw MCMC output from \code{evol.rate.mcmc} and returns a posterior sample of evolutionary rates rootward (\eqn{\sigma_1^2}) and tipward (\eqn{\sigma_2^2}) from the average split.
}
\value{
	A matrix containing the posterior sample of evolutionary rates and shift-points between rates.
}
\references{
	Revell, L. J. (2024) phytools 2.0: an updated R ecosystem for phylogenetic comparative methods (and other things). \emph{PeerJ}, \bold{12}, e16505.

	Revell, L. J., D. L. Mahler, P. Peres-Neto, and B. D. Redelings (2012) A new method for identifying exceptional phenotypic diversification. \emph{Evolution}, \bold{66}, 135-146.
}
\author{Liam Revell \email{liam.revell@umb.edu}}
\seealso{
	\code{\link{evol.rate.mcmc}}, \code{\link{minSplit}}
}
\keyword{phylogenetics}
\keyword{comparative method}
\keyword{bayesian}
\keyword{continuous character}
