\name{print.backbonePhylo}
\alias{print.backbonePhylo}
\title{Print method for backbone phylogeny}
\usage{
\method{print}{backbonePhylo}(x, ...)
}
\arguments{
	\item{x}{an object of class \code{"backbonePhylo"}.}
	\item{...}{optional arguments.}
}
\description{
	Print method for an object of class \code{"backbonePhylo"}.
}
\value{
	Prints to screen.
}
\references{
	Revell, L. J. (2012) phytools: An R package for phylogenetic comparative biology (and other things). \emph{Methods Ecol. Evol.}, \bold{3}, 217-223.
}
\author{Liam Revell \email{liam.revell@umb.edu}}
\seealso{
	\code{\link{phylo.toBackbone}}
}
\keyword{phylogenetics}
\keyword{plotting}

