\name{rstate}
\alias{rstate}
\title{Pick a random state according to a vector of probabilities}
\usage{
rstate(y)
}
\arguments{
  \item{y}{.}
}
\description{
  Internal function for \code{\link{make.simmap}}.
}
\details{
  Do not use unless you know what you're doing.
}
\value{
  a character or string.
}
\references{
}
\author{Liam Revell \email{liam.revell@umb.edu}}
\seealso{
}
\examples{
}
\keyword{phylogenetics}
