\name{splitTree}
\alias{splitTree}
\title{Split tree at a point}
\usage{
splitTree(tree,split)
}
\arguments{
  \item{tree}{.}
  \item{split}{.}
}
\description{
  Internal function for \code{\link{posterior.evolrate}}.
}
\details{
  Do not use unless you know what you're doing.
}
\value{
  two trees in a list.
}
\references{
}
\author{Liam Revell \email{liam.revell@umb.edu}}
\seealso{
}
\examples{
}
\keyword{phylogenetics}
