\name{whichorder}
\alias{whichorder}
\title{Which order are the edges of the tree}
\usage{
whichorder(x,y)
}
\arguments{
  \item{x}{.}
  \item{y}{.}
}
\description{
  Internal function for \code{\link{plotSimmap}}.
}
\details{
  Do not use unless you know what you're doing.
}
\value{
  a vector.
}
\references{
}
\author{Liam Revell \email{liam.revell@umb.edu}}
\seealso{
}
\examples{
}
\keyword{phylogenetics}
