\name{anoletree}
\alias{anoletree}
\title{Phylogeny of Greater Antillean anole ecomorph species with mapped discrete character}
\description{
	A phylogeny of Greater Antillean anole species with a mapped discrete character - 'ecomorph class.' Data and tree are from Mahler et al. (2010).
}
\usage{
data(anoletree)
}
\format{
	The data are stored as a modified object of class \code{"simmap"} with a mapped discrete character. (E.g., see \code{read.simmap}.)
}
\source{
	Mahler, D. L, L. J. Revell, R. E. Glor, and J. B. Losos. (2010) Ecological opportunity and the rate of morphological evolution in the diversification of Greater Antillean anoles. \emph{Evolution}, \bold{64}, 2731-2745.
}
\keyword{datasets}
