\name{as.multiPhylo}
\alias{as.multiPhylo}
\title{Conversion to object of class \code{"multiPhylo"}}
\usage{
as.multiPhylo(x, ...)
\method{as.multiPhylo}{multiSimmap}(x, ...)
}
\arguments{
	\item{x}object to be converted to \code{"multiPhylo"}. Presently an object of class \code{"multiSimmap"}.}
	\item{...}{optional arguments.}
}
\description{
	This function converts between object classes.
}
\value{
	An object of class \code{"multiPhylo"}.
}
\references{
	Revell, L. J. (2012) phytools: An R package for phylogenetic comparative biology (and other things). \emph{Methods Ecol. Evol.}, \bold{3}, 217-223.
}
\author{Liam Revell \email{liam.revell@umb.edu}}
\keyword{phylogenetics}
\keyword{utilities}
