\name{bd}
\alias{bd}
\title{Convert object of class "birthdeath" to raw birth & death rates}
\usage{
bd(x)
}
\arguments{
	\item{x}{object of class \code{"birthdeath"}.}
}
\description{
	This function converts an object of class \code{"birthdeath"} to a vector with the ML birth & death rates.}
}
\value{
	A vector.
}
\references{
	Revell, L. J. (2012) phytools: An R package for phylogenetic comparative biology (and other things). \emph{Methods Ecol. Evol.}, \bold{3}, 217-223.
}
\author{Liam Revell \email{liam.revell@umb.edu}}
\seealso{
	\code{\link{birthdeath}}
}
\keyword{phylogenetics}
\keyword{comparative method}
\keyword{utilities}
