\name{drop.clade}
\alias{drop.clade}
\title{Drop clade from a tree}
\usage{
drop.clade(tree,tip)
}
\arguments{
  \item{tree}{phylogeny.}
  \item{tip}{.}
}
\description{
  Internal function for \code{\link{posterior.evolrate}}.
}
\details{
  Do not use unless you know what you're doing.
}
\value{
  a tree.
}
\references{
}
\author{Liam Revell \email{liam.revell@umb.edu}}
\seealso{
}
\examples{
}
\keyword{phylogenetics}
