\name{expm}
\alias{expm}
\title{Matrix exponential}
\usage{
expm(Y)
}
\arguments{
  \item{Y}{.}
}
\description{
  Internal function for \code{\link{make.simmap}}.
}
\details{
  Do not use unless you know what you're doing.
}
\value{
  a matrix.
}
\references{
}
\author{Liam Revell \email{liam.revell@umb.edu}}
\seealso{
}
\examples{
}
\keyword{phylogenetics}
