\name{expm}
\alias{expm}
\title{Matrix exponential}
\usage{
expm(Y)
}
\arguments{
	\item{Y}{a matrix.}
}
\description{
	Wrapper for \code{\link{MatrixExp}} that retains row/column names.
}
\value{
	A matrix.
}
\references{
	Revell, L. J. (2012) phytools: An R package for phylogenetic comparative biology (and other things). \emph{Methods Ecol. Evol.}, \bold{3}, 217-223.
}
\author{Liam Revell \email{liam.revell@umb.edu}}
\keyword{math}
