\name{gammatest}
\alias{gammatest}
\title{Gamma test of Pybus & Harvey (2000)}
\usage{
gammatest(x)
}
\arguments{
  \item{x}{.}
}
\description{
  Internal function for \code{\link{ltt}}.
}
\details{
  Do not use unless you know what you're doing.
}
\value{
  a tree.
}
\references{
}
\author{Liam Revell \email{liam.revell@umb.edu}}
\seealso{
}
\examples{
}
\keyword{phylogenetics}
