\name{get.treepos}
\alias{get.treepos}
\alias{getnode}
\title{Internal functions}
\usage{
get.treepos(message=TRUE, ...)
getnode(...)
}
\arguments{
	\item{message}{argument.}
	\item{...}{optional arguments.}
}
\description{
	Internally used function.
}
\value{
	A list for \code{get.treepos} and a node number for \code{getnode}.
}
\references{
	Revell, L. J. (2012) phytools: An R package for phylogenetic comparative biology (and other things). \emph{Methods Ecol. Evol.}, \bold{3}, 217-223.
}
\author{Liam Revell \email{liam.revell@umb.edu}}
\keyword{utilities}
