\name{lambda.transform}
\alias{lambda.transform}
\title{Lambda transformation of matrix}
\usage{
lambda.transform(lambda,C)
}
\arguments{
  \item{lambda}{.}
  \item{C}{.}
}
\description{
  Internal function for \code{\link{phyl.pca}}.
}
\details{
  Do not use unless you know what you're doing.
}
\value{
  a matrix.
}
\references{
}
\author{Liam Revell \email{liam.revell@umb.edu}}
\seealso{
}
\examples{
}
\keyword{phylogenetics}
\keyword{math}
\keyword{comparative method}
