\name{paste.tree}
\alias{paste.tree}
\title{Paste tree}
\usage{
paste.tree(tr1,tr2)
}
\arguments{
  \item{tr1}{.}
  \item{tr2}{.}
}
\description{
  Internal function for \code{\link{posterior.evolrate}}.
}
\details{
  Do not use unless you know what you're doing.
}
\value{
  a tree.
}
\references{
}
\author{Liam Revell \email{liam.revell@umb.edu}}
\seealso{
}
\examples{
}
\keyword{phylogenetics}
