\name{phytools-package}
\alias{phytools-package}
\alias{phytools}
\docType{package}
\title{
phytools: Phylogenetic Tools for comparative biology (and other things)
}
\description{
  \pkg{phytools} provides functions for phylogenetic comparative biology; as well as several other functions for tree inference, manipulation, and analysis that are not implemented in other R packages.

  The complete list of functions can be displayed with
  \code{library(help = phytools)}.

  More information on \pkg{phytools} can be found at \url{http://faculty.umb.edu/liam.revell/phytools}.
}
\author{
  Liam J. Revell

  Maintainer: Liam J. Revell <liam.revell@umb.edu>
}
\references{
    Revell, L. J. (In press) phytools: An R package for phylogenetic comparative biology (and other things). Methods Ecol. Evol. doi:10.1111/j.2041-210X.2011.00169.x
}
\keyword{package}
