\name{phytools-package}
\alias{phytools-package}
\alias{phytools}
\docType{package}
\title{
	phytools: Phylogenetic Tools for comparative biology (and other things)
}
\description{
	\pkg{phytools} provides functions for phylogenetic  biology: focused on phylogenetic comparative analysis, but including methods to simulate data and trees, to visualize trees and fitted models, and to analyze and manipulate phylogenetic trees and data.
	
	The complete list of functions can be displayed with \code{library(help=phytools)}.
	
	The \pkg{phytools} development page is \url{https://github.com/liamrevell/phytools/}. More information on \pkg{phytools} can also be found at \url{https://blog.phytools.org} or \url{http://www.phytools.org}.
	
	If you use \pkg{phytools} (or other packages that depend on \pkg{phytools}) in a publication, please \emph{cite it}.
	
	The appropriate citation for \pkg{phytools} is listed below or can be obtained using \code{citation("phytools")} with the package installed.
}
\author{
	Liam J. Revell

	Maintainer: Liam J. Revell <liam.revell@umb.edu>
}
\references{
	Revell, L. J. (2024) phytools 2.0: an updated R ecosystem for phylogenetic comparative methods (and other things). \emph{PeerJ}, \bold{12}, e16505.
}
\keyword{package}
