\name{print.backbonePhylo}
\alias{print.backbonePhylo}
\title{Print method for backbone phylogeny}
\usage{
\method{print}{backbonePhylo}(x, ...)
}
\arguments{
	\item{x}{an object of class \code{"backbonePhylo"}.}
	\item{...}{optional arguments.}
}
\description{
	Print method for an object of class \code{"backbonePhylo"}.
}
\value{
	Prints to screen.
}
\references{
	Revell, L. J. (2024) phytools 2.0: an updated R ecosystem for phylogenetic comparative methods (and other things). \emph{PeerJ}, \bold{12}, e16505.
}
\author{Liam Revell \email{liam.revell@umb.edu}}
\seealso{
	\code{\link{phylo.toBackbone}}
}
\keyword{phylogenetics}
\keyword{plotting}

