\name{setMap}
\alias{setMap}
\title{Set color map for objects of class \code{"contMap"} or \code{"densityMap"}}
\usage{
setMap(x, ...)
}
\arguments{
	\item{x}{an object of class \code{"contMap"} or \code{"densityMap"}.}
	\item{...}{arguments to be passed to \code{\link{colorRampPalette}}.}
}
\description{
	Function to change the color map (ramp) in an object of class \code{"contMap"} or \code{"densityMap"}.
}
\value{
	An object of class \code{"contMap"} or \code{"densityMap"}.
}
\author{Liam Revell \email{liam.revell@umb.edu}}
\seealso{
	\code{\link{contMap}}, \code{\link{densityMap}}
}
\keyword{phylogenetics}
\keyword{plotting}

